requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
joblib==1.2.0
mypy-extensions==1.0.0
numpy==1.23.4
packaging==24.1
pandas==1.5.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
sqlparse==0.5.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.12.2
